Fix NodeManager behaviour; Add tests
ReferenceQueue for Nodes had too little entries. The reason is probably that the NodeRefs (weak references) had been garbage collected before the gc could notice that their referent was unreachable. This is now fixed temporarily by storing all NodeRefs in a list in the manager. This list cannot be resetted reliably without missing entries in the reference queue; Another solution has yet to be found, if we don't want to keep the list of NodeRefs forever.